Problem: there is no way to call ref impl from python
Solution:

Here we make validate.py module and test it with validate_test module.
It creates a temp file, runs JS code against it, deletes it and returns
an instance of ValidatorResult class.

This instance has an optional time spent by the robot and extra information
such as error message.

We also identify venues of improvement of this solution by producing better
error messages in `extra` dict. It is trivial to cover more error messages and
anyone can add some.